Address comments by @trustin / Clean up
