+++ rxjava/src/main/java/com/linecorp/armeria/common/rxjava/RequestContextFlowable.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2020 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,29 +14,37 @@
  * under the License.
  */
 
-package com.linecorp.armeria.rxjava;
+package com.linecorp.armeria.common.rxjava;
+
+import org.reactivestreams.Publisher;
+import org.reactivestreams.Subscriber;
 
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.util.SafeCloseable;
 
-import io.reactivex.rxjava3.core.Observable;
-import io.reactivex.rxjava3.core.ObservableSource;
-import io.reactivex.rxjava3.core.Observer;
+import io.reactivex.rxjava3.core.Flowable;
+import io.reactivex.rxjava3.internal.fuseable.ConditionalSubscriber;
 
-final class RequestContextObservable<T> extends Observable<T> {
+final class RequestContextFlowable<T> extends Flowable<T> {
 
-    private final ObservableSource<T> source;
+    private final Publisher<T> source;
     private final RequestContext assemblyContext;
 
-    RequestContextObservable(ObservableSource<T> source, RequestContext assemblyContext) {
+    RequestContextFlowable(Publisher<T> source, RequestContext assemblyContext) {
         this.source = source;
         this.assemblyContext = assemblyContext;
     }
 
     @Override
-    protected void subscribeActual(Observer<? super T> s) {
+    protected void subscribeActual(Subscriber<? super T> s) {
         try (SafeCloseable ignored = assemblyContext.push()) {
-            source.subscribe(new RequestContextObserver<>(s, assemblyContext));
+            if (s instanceof ConditionalSubscriber) {
+                source.subscribe(new RequestContextConditionalSubscriber<>(
+                        (ConditionalSubscriber<? super T>) s, assemblyContext
+                ));
+            } else {
+                source.subscribe(new RequestContextSubscriber<>(s, assemblyContext));
+            }
         }
     }
 }
