+++ rxjava/src/main/java/com/linecorp/armeria/common/rxjava/RequestContextMaybe.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2020 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.rxjava;
+package com.linecorp.armeria.common.rxjava;
 
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.util.SafeCloseable;
@@ -22,14 +22,13 @@ import com.linecorp.armeria.common.util.SafeCloseable;
 import io.reactivex.rxjava3.core.Maybe;
 import io.reactivex.rxjava3.core.MaybeObserver;
 import io.reactivex.rxjava3.core.MaybeSource;
-import io.reactivex.rxjava3.internal.fuseable.ScalarSupplier;
 
-final class RequestContextScalarSupplierMaybe<T> extends Maybe<T>
-        implements ScalarSupplier<T> {
+final class RequestContextMaybe<T> extends Maybe<T> {
+
     private final MaybeSource<T> source;
     private final RequestContext assemblyContext;
 
-    RequestContextScalarSupplierMaybe(MaybeSource<T> source, RequestContext assemblyContext) {
+    RequestContextMaybe(MaybeSource<T> source, RequestContext assemblyContext) {
         this.source = source;
         this.assemblyContext = assemblyContext;
     }
@@ -38,14 +37,6 @@ final class RequestContextScalarSupplierMaybe<T> extends Maybe<T>
     protected void subscribeActual(MaybeObserver<? super T> s) {
         try (SafeCloseable ignored = assemblyContext.push()) {
             source.subscribe(new RequestContextMaybeObserver<>(s, assemblyContext));
-        }
-    }
-
-    @Override
-    @SuppressWarnings("unchecked")
-    public T get() {
-        try (SafeCloseable ignored = assemblyContext.push()) {
-            return ((ScalarSupplier<T>) source).get();
         }
     }
 }
