+++ rxjava/src/main/java/com/linecorp/armeria/common/rxjava/RequestContextMaybeObserver.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2020 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,66 +14,64 @@
  * under the License.
  */
 
-package com.linecorp.armeria.rxjava;
+package com.linecorp.armeria.common.rxjava;
 
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.util.SafeCloseable;
 
-import io.reactivex.rxjava3.internal.fuseable.ConditionalSubscriber;
-import io.reactivex.rxjava3.internal.fuseable.QueueSubscription;
-import io.reactivex.rxjava3.internal.subscribers.BasicFuseableConditionalSubscriber;
+import io.reactivex.rxjava3.core.MaybeObserver;
+import io.reactivex.rxjava3.disposables.Disposable;
+import io.reactivex.rxjava3.internal.disposables.DisposableHelper;
 
-final class RequestContextConditionalSubscriber<T> extends BasicFuseableConditionalSubscriber<T, T> {
-
+final class RequestContextMaybeObserver<T> implements MaybeObserver<T>, Disposable {
+    private final MaybeObserver<T> actual;
     private final RequestContext assemblyContext;
+    private Disposable disposable;
 
-    RequestContextConditionalSubscriber(ConditionalSubscriber<? super T> downstream,
-                                        RequestContext assemblyContext) {
-        super(downstream);
+    RequestContextMaybeObserver(MaybeObserver<T> actual, RequestContext assemblyContext) {
+        this.actual = actual;
         this.assemblyContext = assemblyContext;
     }
 
     @Override
-    public boolean tryOnNext(T t) {
-        try (SafeCloseable ignored = assemblyContext.push()) {
-            return downstream.tryOnNext(t);
+    public void onSubscribe(Disposable d) {
+        if (!DisposableHelper.validate(disposable, d)) {
+            return;
         }
-    }
-
-    @Override
-    public void onNext(T t) {
+        disposable = d;
         try (SafeCloseable ignored = assemblyContext.push()) {
-            downstream.onNext(t);
+            actual.onSubscribe(this);
         }
     }
 
     @Override
     public void onError(Throwable t) {
         try (SafeCloseable ignored = assemblyContext.push()) {
-            downstream.onError(t);
+            actual.onError(t);
+        }
+    }
+
+    @Override
+    public void onSuccess(T value) {
+        try (SafeCloseable ignored = assemblyContext.push()) {
+            actual.onSuccess(value);
         }
     }
 
     @Override
     public void onComplete() {
         try (SafeCloseable ignored = assemblyContext.push()) {
-            downstream.onComplete();
+            actual.onComplete();
         }
     }
 
     @Override
-    public int requestFusion(int mode) {
-        final QueueSubscription<T> qs = this.qs;
-        if (qs != null) {
-            final int m = qs.requestFusion(mode);
-            sourceMode = m;
-            return m;
-        }
-        return NONE;
+    public boolean isDisposed() {
+        return disposable.isDisposed();
     }
 
     @Override
-    public T poll() throws Throwable {
-        return qs.poll();
+    public void dispose() {
+        disposable.dispose();
     }
 }
