+++ rxjava/src/main/java/com/linecorp/armeria/common/rxjava/RequestContextObserver.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2020 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,42 +14,57 @@
  * under the License.
  */
 
-package com.linecorp.armeria.rxjava;
+package com.linecorp.armeria.common.rxjava;
 
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.util.SafeCloseable;
 
 import io.reactivex.rxjava3.core.Observer;
-import io.reactivex.rxjava3.disposables.Disposable;
-import io.reactivex.rxjava3.functions.Consumer;
-import io.reactivex.rxjava3.observables.ConnectableObservable;
+import io.reactivex.rxjava3.internal.fuseable.QueueDisposable;
+import io.reactivex.rxjava3.internal.observers.BasicFuseableObserver;
 
-final class RequestContextConnectableObservable<T> extends ConnectableObservable<T> {
-
-    private final ConnectableObservable<T> source;
+final class RequestContextObserver<T> extends BasicFuseableObserver<T, T> {
     private final RequestContext assemblyContext;
 
-    RequestContextConnectableObservable(ConnectableObservable<T> source, RequestContext assemblyContext) {
-        this.source = source;
+    RequestContextObserver(Observer<? super T> downstream, RequestContext assemblyContext) {
+        super(downstream);
         this.assemblyContext = assemblyContext;
     }
 
     @Override
-    protected void subscribeActual(Observer<? super T> s) {
+    public void onNext(T t) {
         try (SafeCloseable ignored = assemblyContext.push()) {
-            source.subscribe(new RequestContextObserver<>(s, assemblyContext));
+            downstream.onNext(t);
         }
     }
 
     @Override
-    public void connect(Consumer<? super Disposable> connection) {
+    public void onError(Throwable t) {
         try (SafeCloseable ignored = assemblyContext.push()) {
-            source.connect(connection);
+            downstream.onError(t);
         }
     }
 
     @Override
-    public void reset() {
-        source.reset();
+    public void onComplete() {
+        try (SafeCloseable ignored = assemblyContext.push()) {
+            downstream.onComplete();
+        }
+    }
+
+    @Override
+    public int requestFusion(int mode) {
+        final QueueDisposable<T> qd = this.qd;
+        if (qd != null) {
+            final int m = qd.requestFusion(mode);
+            sourceMode = m;
+            return m;
+        }
+        return NONE;
+    }
+
+    @Override
+    public T poll() throws Throwable {
+        return qd.poll();
     }
 }
