+++ rxjava/src/main/java/com/linecorp/armeria/common/rxjava/RequestContextParallelFlowable.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2020 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.rxjava;
+package com.linecorp.armeria.common.rxjava;
 
 import org.reactivestreams.Subscriber;
 
@@ -50,9 +50,8 @@ final class RequestContextParallelFlowable<T> extends ParallelFlowable<T> {
         for (int i = 0; i < n; i++) {
             final Subscriber<? super T> z = s[i];
             if (z instanceof ConditionalSubscriber) {
-                parents[i] = new RequestContextConditionalSubscriber<>(
-                        (ConditionalSubscriber<? super T>) z, assemblyContext
-                );
+                parents[i] = new RequestContextConditionalSubscriber<>((ConditionalSubscriber<? super T>) z,
+                                                                       assemblyContext);
             } else {
                 parents[i] = new RequestContextSubscriber<>(z, assemblyContext);
             }
