+++ rxjava/src/main/java/com/linecorp/armeria/common/rxjava/RequestContextScalarSupplierMaybe.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2020 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,44 +14,35 @@
  * under the License.
  */
 
-package com.linecorp.armeria.rxjava;
-
-import org.reactivestreams.Publisher;
-import org.reactivestreams.Subscriber;
+package com.linecorp.armeria.common.rxjava;
 
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.util.SafeCloseable;
 
-import io.reactivex.rxjava3.core.Flowable;
-import io.reactivex.rxjava3.internal.fuseable.ConditionalSubscriber;
+import io.reactivex.rxjava3.core.Maybe;
+import io.reactivex.rxjava3.core.MaybeObserver;
+import io.reactivex.rxjava3.core.MaybeSource;
 import io.reactivex.rxjava3.internal.fuseable.ScalarSupplier;
 
-final class RequestContextScalarSupplierFlowable<T> extends Flowable<T>
+final class RequestContextScalarSupplierMaybe<T> extends Maybe<T>
         implements ScalarSupplier<T> {
-    private final Publisher<T> source;
+    private final MaybeSource<T> source;
     private final RequestContext assemblyContext;
 
-    RequestContextScalarSupplierFlowable(Publisher<T> source, RequestContext assemblyContext) {
+    RequestContextScalarSupplierMaybe(MaybeSource<T> source, RequestContext assemblyContext) {
         this.source = source;
         this.assemblyContext = assemblyContext;
     }
 
-    @SuppressWarnings("unchecked")
     @Override
-    protected void subscribeActual(Subscriber<? super T> s) {
+    protected void subscribeActual(MaybeObserver<? super T> s) {
         try (SafeCloseable ignored = assemblyContext.push()) {
-            if (s instanceof ConditionalSubscriber) {
-                source.subscribe(new RequestContextConditionalSubscriber<>(
-                        (ConditionalSubscriber<? super T>) s, assemblyContext
-                ));
-            } else {
-                source.subscribe(new RequestContextSubscriber<>(s, assemblyContext));
-            }
+            source.subscribe(new RequestContextMaybeObserver<>(s, assemblyContext));
         }
     }
 
-    @SuppressWarnings("unchecked")
     @Override
+    @SuppressWarnings("unchecked")
     public T get() {
         try (SafeCloseable ignored = assemblyContext.push()) {
             return ((ScalarSupplier<T>) source).get();
