+++ rxjava/src/main/java/com/linecorp/armeria/common/rxjava/RequestContextScalarSupplierObservable.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2020 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,37 +14,38 @@
  * under the License.
  */
 
-package com.linecorp.armeria.rxjava;
-
-import org.reactivestreams.Publisher;
-import org.reactivestreams.Subscriber;
+package com.linecorp.armeria.common.rxjava;
 
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.util.SafeCloseable;
 
-import io.reactivex.rxjava3.core.Flowable;
-import io.reactivex.rxjava3.internal.fuseable.ConditionalSubscriber;
+import io.reactivex.rxjava3.core.Observable;
+import io.reactivex.rxjava3.core.ObservableSource;
+import io.reactivex.rxjava3.core.Observer;
+import io.reactivex.rxjava3.internal.fuseable.ScalarSupplier;
 
-final class RequestContextFlowable<T> extends Flowable<T> {
+final class RequestContextScalarSupplierObservable<T> extends Observable<T> implements ScalarSupplier<T> {
 
-    private final Publisher<T> source;
+    private final ObservableSource<T> source;
     private final RequestContext assemblyContext;
 
-    RequestContextFlowable(Publisher<T> source, RequestContext assemblyContext) {
+    RequestContextScalarSupplierObservable(ObservableSource<T> source, RequestContext assemblyContext) {
         this.source = source;
         this.assemblyContext = assemblyContext;
     }
 
     @Override
-    protected void subscribeActual(Subscriber<? super T> s) {
+    protected void subscribeActual(Observer<? super T> s) {
         try (SafeCloseable ignored = assemblyContext.push()) {
-            if (s instanceof ConditionalSubscriber) {
-                source.subscribe(new RequestContextConditionalSubscriber<>(
-                        (ConditionalSubscriber<? super T>) s, assemblyContext
-                ));
-            } else {
-                source.subscribe(new RequestContextSubscriber<>(s, assemblyContext));
-            }
+            source.subscribe(new RequestContextObserver<>(s, assemblyContext));
+        }
+    }
+
+    @SuppressWarnings("unchecked")
+    @Override
+    public T get() {
+        try (SafeCloseable ignored = assemblyContext.push()) {
+            return ((ScalarSupplier<T>) source).get();
         }
     }
 }
