+++ rxjava/src/main/java/com/linecorp/armeria/common/rxjava/RequestContextScalarSupplierSingle.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2020 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,30 +14,30 @@
  * under the License.
  */
 
-package com.linecorp.armeria.rxjava;
+package com.linecorp.armeria.common.rxjava;
 
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.util.SafeCloseable;
 
-import io.reactivex.rxjava3.core.Observable;
-import io.reactivex.rxjava3.core.ObservableSource;
-import io.reactivex.rxjava3.core.Observer;
+import io.reactivex.rxjava3.core.Single;
+import io.reactivex.rxjava3.core.SingleObserver;
+import io.reactivex.rxjava3.core.SingleSource;
 import io.reactivex.rxjava3.internal.fuseable.ScalarSupplier;
 
-final class RequestContextScalarSupplierObservable<T> extends Observable<T> implements ScalarSupplier<T> {
-
-    private final ObservableSource<T> source;
+final class RequestContextScalarSupplierSingle<T> extends Single<T>
+        implements ScalarSupplier<T> {
+    private final SingleSource<T> source;
     private final RequestContext assemblyContext;
 
-    RequestContextScalarSupplierObservable(ObservableSource<T> source, RequestContext assemblyContext) {
+    RequestContextScalarSupplierSingle(SingleSource<T> source, RequestContext assemblyContext) {
         this.source = source;
         this.assemblyContext = assemblyContext;
     }
 
     @Override
-    protected void subscribeActual(Observer<? super T> s) {
+    protected void subscribeActual(SingleObserver<? super T> s) {
         try (SafeCloseable ignored = assemblyContext.push()) {
-            source.subscribe(new RequestContextObserver<>(s, assemblyContext));
+            source.subscribe(new RequestContextSingleObserver<>(s, assemblyContext));
         }
     }
