+++ rxjava/src/main/java/com/linecorp/armeria/internal/server/rxjava/ObservableResponseConverterFunctionProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2020 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
